+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 1.3 (unreleased)
 ----------------
 
+- Add ajax_load=1 to actions entries (@@sharing, folder_contents, etc.) for
+  faster loading without the plone border.
+  [thet]
+
 - Fix ZCML loading using z3c.autoinclude.
   [rpatterson]